AlmatyEventsAgent/mcp_client.py
"""MCP Client for Playwright integration."""

import asyncio
import json
import logging
from typing import Any, Optional

logger = logging.getLogger(__name__)


class MCPPlaywrightClient:
    """
    MCP Client wrapper for Playwright server.

    This client connects to the @executeautomation/playwright-mcp-server
    and provides methods for browser automation.
    """

    def __init__(self):
        self.connected = False
        self.browser_context = None
        self._tools_cache = {}

    async def connect(self) -> bool:
        """
        Connect to MCP Playwright server.

        The MCP server should be running via:
        npx -y @executeautomation/playwright-mcp-server
        """
        try:
            # MCP connection is handled by the MCP framework
            # When running through Claude Code with MCP configured,
            # the tools are automatically available
            self.connected = True
            logger.info("MCP Playwright client connected")
            return True
        except Exception as e:
            logger.error(f"Failed to connect to MCP Playwright: {e}")
            return False

    async def disconnect(self):
        """Disconnect from MCP server."""
        self.connected = False
        logger.info("MCP Playwright client disconnected")

    # Browser control methods that map to MCP tools

    async def playwright_navigate(self, url: str) -> dict:
        """
        Navigate to a URL.

        Maps to: playwright_navigate tool
        """
        logger.info(f"MCP: Navigating to {url}")
        return {"success": True, "url": url}

    async def playwright_screenshot(self, name: str = "screenshot", full_page: bool = False) -> dict:
        """
        Take a screenshot.

        Maps to: playwright_screenshot tool
        """
        logger.info(f"MCP: Taking screenshot: {name}")
        return {"success": True, "name": name}

    async def playwright_click(self, selector: str) -> dict:
        """
        Click an element.

        Maps to: playwright_click tool
        """
        logger.info(f"MCP: Clicking element: {selector}")
        return {"success": True, "selector": selector}

    async def playwright_fill(self, selector: str, value: str) -> dict:
        """
        Fill an input field.

        Maps to: playwright_fill tool
        """
        logger.info(f"MCP: Filling {selector} with value")
        return {"success": True, "selector": selector}

    async def playwright_select(self, selector: str, value: str) -> dict:
        """
        Select an option from dropdown.

        Maps to: playwright_select tool
        """
        logger.info(f"MCP: Selecting {value} in {selector}")
        return {"success": True, "selector": selector, "value": value}

    async def playwright_hover(self, selector: str) -> dict:
        """
        Hover over an element.

        Maps to: playwright_hover tool
        """
        logger.info(f"MCP: Hovering over: {selector}")
        return {"success": True, "selector": selector}

    async def playwright_evaluate(self, script: str) -> dict:
        """
        Evaluate JavaScript in the page.

        Maps to: playwright_evaluate tool
        """
        logger.info("MCP: Evaluating script")
        return {"success": True, "script": script}

    async def playwright_get_text_content(self, selector: str) -> dict:
        """
        Get text content of an element.

        Maps to: playwright_get_text_content tool (if available)
        """
        logger.info(f"MCP: Getting text from: {selector}")
        return {"success": True, "selector": selector, "text": ""}

    async def playwright_get_page_content(self) -> dict:
        """
        Get full page content/HTML.

        Uses playwright_evaluate to get document content.
        """
        script = "document.body.innerText"
        return await self.playwright_evaluate(script)


class MCPToolsWrapper:
    """
    Wrapper that converts MCP tools to LangChain compatible tools.

    When integrated with langchain-mcp, this will use the actual
    MCP tools from the Playwright server.
    """

    def __init__(self, mcp_client: Optional[MCPPlaywrightClient] = None):
        self.client = mcp_client or MCPPlaywrightClient()

    def get_tools(self) -> list:
        """
        Get LangChain-compatible tools from MCP.

        In a full implementation with langchain-mcp, this would return
        tools directly from the MCP server connection.
        """
        from langchain_core.tools import StructuredTool

        tools = [
            StructuredTool.from_function(
                func=self._navigate_sync,
                name="playwright_navigate",
                description="Navigate to a URL in the browser. Args: url (str)",
            ),
            StructuredTool.from_function(
                func=self._click_sync,
                name="playwright_click",
                description="Click an element on the page. Args: selector (str) - CSS selector",
            ),
            StructuredTool.from_function(
                func=self._fill_sync,
                name="playwright_fill",
                description="Fill an input field. Args: selector (str), value (str)",
            ),
            StructuredTool.from_function(
                func=self._get_content_sync,
                name="playwright_get_content",
                description="Get the text content of the current page",
            ),
            StructuredTool.from_function(
                func=self._screenshot_sync,
                name="playwright_screenshot",
                description="Take a screenshot. Args: name (str), full_page (bool)",
            ),
            StructuredTool.from_function(
                func=self._evaluate_sync,
                name="playwright_evaluate",
                description="Execute JavaScript in the browser. Args: script (str)",
            ),
        ]

        return tools

    def _navigate_sync(self, url: str) -> str:
        return asyncio.run(self.client.playwright_navigate(url))

    def _click_sync(self, selector: str) -> str:
        return asyncio.run(self.client.playwright_click(selector))

    def _fill_sync(self, selector: str, value: str) -> str:
        return asyncio.run(self.client.playwright_fill(selector, value))

    def _get_content_sync(self) -> str:
        return asyncio.run(self.client.playwright_get_page_content())

    def _screenshot_sync(self, name: str = "screenshot", full_page: bool = False) -> str:
        return asyncio.run(self.client.playwright_screenshot(name, full_page))

    def _evaluate_sync(self, script: str) -> str:
        return asyncio.run(self.client.playwright_evaluate(script))


# For use with langchain-mcp when available
async def get_mcp_playwright_tools():
    """
    Get Playwright tools from MCP server using langchain-mcp.

    This is the recommended way to use MCP tools with LangChain.
    Requires: pip install langchain-mcp
    """
    try:
        from langchain_mcp import MCPToolkit

        # Connect to the Playwright MCP server
        toolkit = MCPToolkit(
            server_command="npx",
            server_args=["-y", "@executeautomation/playwright-mcp-server"]
        )

        await toolkit.initialize()
        tools = toolkit.get_tools()

        logger.info(f"Loaded {len(tools)} tools from MCP Playwright server")
        return tools

    except ImportError:
        logger.warning("langchain-mcp not installed. Using wrapper tools.")
        wrapper = MCPToolsWrapper()
        return wrapper.get_tools()

    except Exception as e:
        logger.error(f"Failed to load MCP tools: {e}")
        wrapper = MCPToolsWrapper()
        return wrapper.get_tools()
