AlmatyEventsAgent/requirements.txt
# AlmatyEventsAgent Dependencies

# Core LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# LangGraph for state management
langgraph>=0.2.0

# MCP Integration
# langchain-mcp>=0.1.0  # Uncomment when available

# Async support
aiohttp>=3.9.0

# Environment and config
python-dotenv>=1.0.0
pydantic>=2.0.0

# CLI and formatting
rich>=13.0.0

# Playwright (for browser automation)
playwright>=1.40.0

# Utilities
httpx>=0.25.0

# Optional: For web scraping fallback
beautifulsoup4>=4.12.0
lxml>=5.0.0
